app/requirements.txt
fastapi==0.121.0
httpx==0.28.1
joblib==1.5.2
nameparser==1.1.3
python_dateutil==2.9.0.post0
scikit_learn==1.7.2
spacy==3.8.7
uvicorn[standard]==0.38.0
python-multipart==0.0.20
